# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Classroom management backend (Go net/http ServeMux). Auth via JWT Bearer (HS256, 24h); mini-program identity resolved server-side via code2Session.",
        "title": "OCM 智慧教室管理平台 API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/dashboard/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Each section is present only when the caller holds the\npermission that gates it — absence means \"hidden\".",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Console homepage summary (permission-gated sections)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "school day (Y-M-D)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sectioned summary",
                        "schema": {
                            "$ref": "#/definitions/dashboard.Summary"
                        }
                    },
                    "400": {
                        "description": "date must be YYYY-MM-DD",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "not authenticated",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/settings/email": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get mail service settings (password masked)",
                "responses": {
                    "200": {
                        "description": "settings with the password hidden",
                        "schema": {
                            "$ref": "#/definitions/mail.MaskedSettings"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "An empty password means \"keep the stored one\" — the secret\ncan never be cleared through the API.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update mail service settings",
                "parameters": [
                    {
                        "description": "mail settings",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mail.Settings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated settings (password masked)",
                        "schema": {
                            "$ref": "#/definitions/mail.MaskedSettings"
                        }
                    },
                    "400": {
                        "description": "invalid body / required fields missing when enabled",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "booking.BookingView": {
            "type": "object",
            "required": [
                "classroomId",
                "classroomName",
                "createdAt",
                "date",
                "displayName",
                "id",
                "periodEnd",
                "periodStart",
                "purpose",
                "status",
                "userId",
                "username"
            ],
            "properties": {
                "classroomId": {
                    "type": "integer"
                },
                "classroomName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "description": "\"YYYY-MM-DD\"",
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "periodEnd": {
                    "type": "integer"
                },
                "periodStart": {
                    "type": "integer"
                },
                "purpose": {
                    "type": "string"
                },
                "reviewedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "booking.DailyCount": {
            "type": "object",
            "required": [
                "count",
                "date"
            ],
            "properties": {
                "count": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                }
            }
        },
...
        "classroom.RepairView": {
            "type": "object",
            "required": [
                "assigneeName",
                "building",
                "classroomId",
                "classroomName",
                "createdAt",
                "creatorId",
                "creatorName",
                "description",
                "id",
                "images",
                "remark",
                "status",
                "updatedAt"
            ],
            "properties": {
                "assigneeId": {
                    "type": "integer"
                },
                "assigneeName": {
                    "type": "string"
                },
                "building": {
                    "type": "string"
                },
                "classroomId": {
                    "type": "integer"
                },
                "classroomName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "creatorId": {
                    "type": "integer"
                },
                "creatorName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "course.SessionView": {
            "type": "object",
            "required": [
                "catalogCode",
                "classNames",
                "classroomId",
                "classroomName",
                "courseName",
                "createdAt",
                "date",
                "id",
                "note",
                "offeringId",
                "periodEnd",
                "periodStart",
                "semester",
                "teacher",
                "teachingClassName"
            ],
            "properties": {
                "catalogCode": {
                    "type": "string"
                },
                "classNames": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "classroomId": {
                    "type": "integer"
                },
                "classroomName": {
                    "type": "string"
                },
                "courseName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "description": "\"YYYY-MM-DD\"",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "offeringId": {
                    "type": "integer"
                },
                "periodEnd": {
                    "type": "integer"
                },
                "periodStart": {
                    "type": "integer"
                },
                "semester": {
                    "type": "string"
                },
                "teacher": {
                    "type": "string"
                },
                "teachingClassName": {
                    "type": "string"
                }
            }
        },
...
        "dashboard.BookingsSection": {
            "type": "object",
            "required": [
                "items",
                "total"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/booking.BookingView"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dashboard.PeriodCount": {
            "type": "object",
            "required": [
                "count",
                "period"
            ],
            "properties": {
                "count": {
                    "type": "integer"
                },
                "period": {
                    "type": "integer"
                }
            }
        },
        "dashboard.RepairsSection": {
            "type": "object",
            "required": [
                "items",
                "total"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/classroom.RepairView"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dashboard.SessionsSection": {
            "type": "object",
            "required": [
                "items",
                "total"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/course.SessionView"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dashboard.Summary": {
            "type": "object",
            "required": [
                "date"
            ],
            "properties": {
                "bookingLoad": {
                    "description": "BookingLoad is approved booking counts per day over the next 14 days,\nzero-filled. Scope matches the caller: booking:approve sees every\nbooking, everyone else only their own. Visible to every authenticated\nsubject.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/booking.DailyCount"
                    }
                },
                "classroomTotal": {
                    "description": "ClassroomTotal is the number of managed classrooms (classroom:read).",
                    "type": "integer"
                },
                "date": {
                    "description": "Date is the echoed \"YYYY-MM-DD\" school day the figures refer to\n(supplied by the client so server/client timezone skew cannot shift\n\"today\").",
                    "type": "string"
                },
                "myBookings": {
                    "description": "MyBookings is the caller's own approved bookings starting at Date,\nsoonest first. Visible to every authenticated subject: a booking's\ncreator can always see their own reservation.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/booking.BookingView"
                    }
                },
                "openRepairs": {
                    "description": "OpenRepairs is the unresolved ticket queue, open or processing,\nscoped like the repair list itself: repair:assign sees every ticket,\na plain repair:create only their own (repair:create OR repair:assign).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dashboard.RepairsSection"
                        }
                    ]
                },
                "pendingBookings": {
                    "description": "PendingBookings is the pending-approval queue (booking:approve).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dashboard.BookingsSection"
                        }
                    ]
                },
                "recentLogs": {
                    "description": "RecentLogs is the latest audit trail entries (log:read).",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/systemlog.LogView"
                    }
                },
                "sessionPeriods": {
                    "description": "SessionPeriods is today's occupancy histogram: for each period index\n1..max, how many sessions are running (a session spanning periods 1-2\ncounts in both slots). Filled alongside TodaySessions from the same\nquery, so it carries the same course:read gate (course:read).",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dashboard.PeriodCount"
                    }
                },
                "todaySessions": {
                    "description": "TodaySessions is the day's course session count plus a preview of the\nfirst sessions by period (course:read).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dashboard.SessionsSection"
                        }
                    ]
                }
            }
        },
        "httpx.ErrorResponse": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
...
        "mail.MaskedSettings": {
            "type": "object",
            "required": [
                "enabled",
                "encryption",
                "fromAddress",
                "fromName",
                "host",
                "password",
                "passwordSet",
                "port",
                "username"
            ],
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "encryption": {
                    "type": "string"
                },
                "fromAddress": {
                    "type": "string"
                },
                "fromName": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "passwordSet": {
                    "type": "boolean"
                },
                "port": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "mail.Settings": {
            "type": "object",
            "required": [
                "enabled",
                "encryption",
                "fromAddress",
                "fromName",
                "host",
                "password",
                "port",
                "username"
            ],
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "encryption": {
                    "type": "string"
                },
                "fromAddress": {
                    "type": "string"
                },
                "fromName": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "systemlog.LogView": {
            "type": "object",
            "required": [
                "actorName",
                "clientIp",
                "createdAt",
                "id",
                "method",
                "path",
                "statusCode",
                "summary"
            ],
            "properties": {
                "actorId": {
                    "description": "null when the row has no actor",
                    "type": "integer"
                },
                "actorName": {
                    "type": "string"
                },
                "clientIp": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"JWT Bearer token; prefix with 'Bearer '.\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}